backEnd: deleting limits

+++ src/groups/group.controller.ts
@@ -5,6 +5,8 @@ import NotFoundException from "../exceptions/NotFoundException";
 import HttpException from "../exceptions/HttpException";
 import validationMiddleware from "../middleware/validation.middleware";
 import CreateGroupDto from "./group.dto";
+import IdGroupDto from "./idGroup.dto";
+import User from "../users/user.entity";
 
 
 class GroupController {
@@ -16,13 +18,14 @@ class GroupController {
     }
 
     private groupRepository = getRepository(Group)
+    private userRepository = getRepository(User)
 
     public intializeRoutes() {
         this.router.get(this.path, this.getAll);
-        this.router.get('/group', validationMiddleware(CreateGroupDto), this.findOne);
+        this.router.get('/group', validationMiddleware(IdGroupDto), this.findOne);
         this.router.post('/update', validationMiddleware(CreateGroupDto), this.update);
-        this.router.post('/del', validationMiddleware(CreateGroupDto), this.delete);
-        this.router.post('/create', validationMiddleware(CreateGroupDto),  this.create);
+        this.router.post('/del', validationMiddleware(IdGroupDto), this.delete);
+        this.router.post('/create', validationMiddleware(CreateGroupDto), this.create);
     }
 
     private create = async (req: express.Request, res: express.Response, next: express.NextFunction) => {
@@ -33,7 +36,7 @@ class GroupController {
             group.description = description;
             const createdGroup = await this.groupRepository.save(group);
             return res.json(createdGroup)
-        }catch (e) {
+        } catch (e) {
             next(new HttpException(404, e.message));
         }
 
@@ -47,7 +50,7 @@ class GroupController {
             } else {
                 next(new NotFoundException(String(id)));
             }
-        }catch (e) {
+        } catch (e) {
             next(new HttpException(404, e.message));
         }
 
@@ -56,9 +59,14 @@ class GroupController {
         try {
             const {id} = req.body
             if (id) {
+                const groupUsers = await this.userRepository.count({groupId:id});
+                if (groupUsers > 0) {
+                    return res.json('Group is not empty, it counts: ' + groupUsers + ' users!')
+                }
                 let groupToRemove = await this.groupRepository.findOne({id})
                 await this.groupRepository.remove(groupToRemove)
                 return res.json('deleted: ' + groupToRemove.name)
+
             } else {
                 next(new NotFoundException(String(id)));
             }
@@ -77,11 +85,11 @@ class GroupController {
                 groupToUpdate.description = description;
                 let updatedGroup = await this.groupRepository.save(groupToUpdate)
                 return res.json(updatedGroup)
-            }else {
+            } else {
                 next(new NotFoundException(String(id)));
             }
 
-        }catch (e) {
+        } catch (e) {
             next(new HttpException(404, e.message));
         }
 
@@ -91,7 +99,7 @@ class GroupController {
         try {
             const groups = await this.groupRepository.find();
             return res.json(groups)
-        }catch (e) {
+        } catch (e) {
             next(new HttpException(404, e.message));
         }
     }
